Add apt-get imagemagick to dockerfile
In order for pycortex to be able to render SVG-based regions of interest on flatmaps, it turns out that we need the imagemagick `convert` binary installed. I added a line to install all of imagemagick. 

Is there any chance for this to be staged + added to production by tomorrow? (That is the date of our last cogneuro connector class and it would be amazing if we could have this functionality for it).

Thanks!
Michael