w261: Create GCP resources for users
When a user pod starts, we ensure the following GCP
resources:

1. A GCS bucket
2. A Service Account with admin rights on the bucket
3. A Service Account Key for the Service Account

We then populate two environment variables in the
spawned pod

1. SPARK_GCS_BUCKET - name of GCS bucket for this user
2. SPARK_GCS_KEY - JSON Private key for service account

This is all done through a custom Spawner that inherits
from KubeSpawner. This should operate at the level of
a 'Spark+GCS on GCP', so nothing too UCB specific.

In addition, there's a custom role created in the
ucb-datahub-2018 project - projects/ucb-datahub-2018/roles/w261_hub_resource_creator
that has the appropriate permissions needed by the *hub*
to create the resources it needs.