Deploy unmgerged z2jh changes to get new HTTPS
This is madness, but I want to deploy https://github.com/jupyterhub/zero-to-jupyterhub-k8s/pull/1539
before it is merged. So we add that in as a git submodule,
run chartpress without building images (since these images are
built already and pushed by me), and deploy from here.

Forgive me father for I have sinned.