Move *just* data100 to a new 'beta' pool
The data100 image is about 10G. When a new node spawns, it
must pull in this image, in addition to the datahub image (which
is already about 9G) before stuff can realistically start. This
has lead to extremely long node spin up times.

We're going to split data100 into its own node pool, with its
own autoscaling behavior. This should help reduce the total
size of images on each node, and parallelize the scale-up behavior

stat89a did not have a specific node selector, so we add one
to make it use the default node pool