Self-shutdown notebooks after 1h of no activity
The culler doesn't kill some user pods when it should. We now
get the notebook server to try and stop itself when it's got
no activity either. Not sure if this will help, but let's
try! This is what mybinder.org does.

See https://github.com/jupyterhub/jupyterhub/issues/3099
for more info.