Provide separate .rstudio for datahub & r hub
RStudio can't write session files to anywhere except ~/.rstudio -
only way to change that is by setting $HOME. So instead, we just
bind mount a fresh directory on top of ~/.rstudio!

https://github.com/berkeley-dsep-infra/datahub/issues/1899 is
most likely due to two different RStudio versions with
different versions of R writing to the same directory with a
binary serializations from datahub & R hub. We already made the
version the same - but by giving each their own .rstudio, it will
hopefully reduce the likelihood of this.

Ref #1899. Supersedes #1922